Added ability to pull in slides from gists
Note that in order to do so and have things stay in order, I ended
up having to remove the async calls to readFile and make them
readFileSync instead.  Will make things slightly slower when
building from lots of disk files (and is "un-node-like"), but for
quick-and-dirty inclusion of gists it works.